+++ ado_wrapper/__init__.py
@@ -2,14 +2,12 @@
 from ado_wrapper.client import AdoClient
 from ado_wrapper.resources import *
 
-# from ado_wrapper.errors import *
-
 __all__ = [
     "AdoClient",
     "AgentPool", "AnnotatedTag", "Artifact", "AuditLog", "Branch", "BuildTimeline", "Build", "BuildDefinition", "Commit",
     "Environment", "PipelineAuthorisation", "Group", "HierarchyCreatedBuildDefinition", "MergeBranchPolicy", "MergePolicies",
     "MergePolicyDefaultReviewer", "MergeTypeRestrictionPolicy", "Organisation", "Permission", "PersonalAccessToken", "Project",
-    "ProjectRepositorySettings", "PullRequest", "Release", "ReleaseDefinition", "RepoUserPermissions", "UserPermission",
+    "PullRequest", "Release", "ReleaseDefinition", "RepoUserPermissions", "UserPermission",
     "BuildRepository", "Repo", "Run", "CodeSearch", "ServiceEndpoint", "Team", "AdoUser", "Member", "Reviewer", "TeamMember",
     "VariableGroup",
 ]  # fmt: skip