+++ ado_wrapper/resources/__init__.py
@@ -14,7 +14,6 @@ from ado_wrapper.resources.organisations import Organisation
 from ado_wrapper.resources.permissions import Permission
 from ado_wrapper.resources.personal_access_tokens import PersonalAccessToken
 from ado_wrapper.resources.projects import Project
-from ado_wrapper.resources.project_settings import ProjectRepositorySettings, ProjectRepositoryPolicies
 from ado_wrapper.resources.pull_requests import PullRequest
 from ado_wrapper.resources.releases import Release, ReleaseDefinition
 from ado_wrapper.resources.repo_user_permission import RepoUserPermissions, UserPermission
@@ -29,7 +28,7 @@ from ado_wrapper.resources.variable_groups import VariableGroup
 __all__ = [
     "AgentPool", "AnnotatedTag", "Artifact", "AuditLog", "Branch", "BuildTimeline", "Build", "BuildDefinition", "HierarchyCreatedBuildDefinition",
     "Commit", "Environment", "PipelineAuthorisation", "Group", "MergeBranchPolicy", "MergePolicies", "MergePolicyDefaultReviewer",
-    "MergeTypeRestrictionPolicy", "Organisation", "Permission", "PersonalAccessToken", "Project", "ProjectRepositorySettings", "ProjectRepositoryPolicies",
+    "MergeTypeRestrictionPolicy", "Organisation", "Permission", "PersonalAccessToken", "Project",
     "PullRequest", "Release", "ReleaseDefinition", "RepoUserPermissions", "UserPermission", "BuildRepository", "Repo", "Run", "CodeSearch",
     "ServiceEndpoint", "Team", "AdoUser", "Member", "Reviewer", "TeamMember", "VariableGroup"  # fmt: skip
 ]